+++ noodles-gtf/src/lib.rs
@@ -6,9 +6,3 @@ mod line_buf;
 pub mod record;
 
 pub use self::{line::Line, line_buf::LineBuf, record::Record};
-
-#[deprecated(since = "0.32.0", note = "Use `noodles_gtf::io::Reader` instead.")]
-pub use self::io::Reader;
-
-#[deprecated(since = "0.32.0", note = "Use `noodles_gtf::io::Writer` instead.")]
-pub use self::io::Writer;